Split ram resources out of the ram module to avoid conflicts when building go
